Hannibal suggested changes
