+++ src/main/java/at/hannibal2/skyhanni/features/combat/cocoon/CocoonOverlay.kt
@@ -3,10 +3,10 @@ package at.hannibal2.skyhanni.features.combat.cocoon
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.api.event.HandleEvent
 import at.hannibal2.skyhanni.events.minecraft.SkyHanniRenderWorldEvent
+import at.hannibal2.skyhanni.features.combat.cocoon.CocoonAPI.canCocoon
 import at.hannibal2.skyhanni.features.combat.cocoon.CocoonAPI.existingCocoons
 import at.hannibal2.skyhanni.features.combat.cocoon.CocoonAPI.expectedLifetime
 import at.hannibal2.skyhanni.skyhannimodule.SkyHanniModule
-import at.hannibal2.skyhanni.utils.CocoonUtils.canCocoon
 import at.hannibal2.skyhanni.utils.ColorUtils
 import at.hannibal2.skyhanni.utils.ExtendedChatColor
 import at.hannibal2.skyhanni.utils.LocationUtils.distanceToPlayer
@@ -19,12 +19,11 @@ import at.hannibal2.skyhanni.utils.render.WorldRenderUtils.drawDynamicText
 object CocoonOverlay {
     private val config get() = SkyHanniMod.feature.combat.cocoonOverlay
 
-
     @HandleEvent
     fun onRenderWorld(event: SkyHanniRenderWorldEvent) {
+        if (!canCocoon && config.onlyShowWhenPlayerIsAbleToCocoon) return
         for (cocoon in existingCocoons) {
             if (!cocoon.hasBeenSeen && cocoon.coordinates.distanceToPlayer() > 32) continue
-            if (!canCocoon && config.onlyShowWhenPlayerIsAbleToCocoon) return
             if (config.showCocoonContainedMobName) event.drawDynamicText(
                 cocoon.coordinates,
                 cocoon.mob.name,
